+++ core/src/main/java/com/adobe/aem/commons/assetshare/configuration/package-info.java
@@ -17,7 +17,7 @@
  *
  */
 
-@Version("1.2.0")
+@Version("1.3.0")
 package com.adobe.aem.commons.assetshare.configuration;
 
 import org.osgi.annotation.versioning.Version;