+++ ui.apps/src/main/content/jcr_root/apps/asset-share-commons/clientlibs/clientlib-site/js/form-data.js
@@ -21,10 +21,17 @@
 AssetShare.FormData = function (formEl) {
     "use strict";
 
-    var form = [];
+    var form = [],
+        formId;
 
     if (formEl) {
-        form = $(formEl).serializeArray();
+        formId = formEl.attr('id');
+    }
+
+    if (formId) {
+        // Select everything using the form attribute manually for IE11 support
+        // But ONLY if the formEl exists; otherwise leave empty array
+        form = $('[form="' + formId + '"]').serializeArray();
     }
 
     /* IE polyfills for findIndex() and find() */
@@ -42,13 +49,13 @@ AssetShare.FormData = function (formEl) {
     }
 
     function find(key) {
-       var index = findIndex(key);
+        var index = findIndex(key);
 
-       if (index > -1) {
-           return form[index];
-       }
+        if (index > -1) {
+            return form[index];
+        }
 
-       return null;
+        return null;
     }
 
     function forEach(fn) {