+++ CHANGELOG.md
@@ -4,6 +4,27 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [v1.9.0]
+
+### Added
+
+- 0428: Jenkins pipeline for automating releases.
+- 0432: Added the External Redirection Asset Rendition Dispatcher - including updated UrlUrl.escape() support for more complex URLs.
+- 0448: Added AEM Content Services Sling Model Exporter annotations to Asset Share Commons WCM component Sling Models.
+
+### Changed
+
+- 0453: Added support to the Asset Details - Renditions component for the new ASC Asset Rendition framework
+
+### Fixed
+
+- 0430: Fixes issue with default search result layout is incorrect when Statistics component is added above the search results component.
+- 0435: Fixes issue with Dropdown rendering of Search components: Property, Tag and Paths.
+- 0443: Fixed issue on Search Page authoring where page-breaking errors are thrown if a Search Results component has not been added yet (Sort, Filter Toggle and Statistics).
+- 0458: Fixed issue with Form Submissions in IE11
+- 0467: Fixed issue in removing asset in cart modal with large number of assets.
+- 0472: Fixed issue with InternalRedirectRenditionDispatcherImpl failing to dispatch when /etc/map is configured
+
 ## [v1.8.0]
 
 ### Added